+++ app/src/main/cpp/Libraries/2D/GR/grilin.c
@@ -68,7 +68,7 @@ int32_t gr_int_line (int16_t x0, int16_t y0, int16_t x1, int16_t y1)
     return r;
 }
 
-int32_t gri_iline_clip_fill (int32_t c, int32_t parm, grs_vertex *v0, grs_vertex *v1)
+int32_t gri_iline_clip_fill (int32_t c, uintptr_t parm, grs_vertex *v0, grs_vertex *v1)
 {
     int32_t r;
     grs_vertex u0, u1;
